refactor: move port mapping implementation to controller pkg
