+++ pkg/controller/portmapping.go
@@ -1,4 +1,4 @@
-package k8s
+package controller
 
 import (
 	"errors"
@@ -74,7 +74,7 @@ func (p *PortMapping) LoadState() error {
 }
 
 // Find searches for the port which is associated with the given ServiceWithPort.
-func (p *PortMapping) Find(svc ServiceWithPort) (int32, bool) {
+func (p *PortMapping) Find(svc k8s.ServiceWithPort) (int32, bool) {
 	p.mu.RLock()
 	defer p.mu.RUnlock()
 
@@ -89,7 +89,7 @@ func (p *PortMapping) Find(svc ServiceWithPort) (int32, bool) {
 
 // Add adds a new mapping between the given ServiceWithPort and the first port available in the range defined
 // within minPort and maxPort. If there's no port left, an error will be returned.
-func (p *PortMapping) Add(svc *ServiceWithPort) (int32, error) {
+func (p *PortMapping) Add(svc *k8s.ServiceWithPort) (int32, error) {
 	for i := p.minPort; i < p.maxPort+1; i++ {
 		// Skip until an available port is found
 		if _, exists := p.table[i]; exists {
@@ -107,7 +107,7 @@ func (p *PortMapping) Add(svc *ServiceWithPort) (int32, error) {
 }
 
 // Remove removes the mapping associated with the given ServiceWithPort.
-func (p *PortMapping) Remove(svc ServiceWithPort) (int32, error) {
+func (p *PortMapping) Remove(svc k8s.ServiceWithPort) (int32, error) {
 	port, ok := p.Find(svc)
 	if !ok {
 		return 0, fmt.Errorf("unable to find port mapping for service %s/%s on port %d", svc.Namespace, svc.Name, svc.Port)
